// react/src/__rust_generated__/Format.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Format =
    | { type: 'Automatic' }
    | { type: 'PlainText' }
    | { type: 'Money_v2'; precision: number }
    | { type: 'Percent_v2'; precision: number }
    | { type: 'Number'; precision: number }
    | { type: 'Date' }
    | { type: 'DateTime' }
    | { type: 'Money' }
    | { type: 'Percent' };